修复 OracleAI 上下文上限保存

+++ backend/src/main/java/com/ociworker/model/entity/OciOpenaiLbMember.java
@@ -30,7 +30,7 @@ public class OciOpenaiLbMember {
     @TableField(updateStrategy = FieldStrategy.ALWAYS)
     private Long tpmLimit;
     @TableField(updateStrategy = FieldStrategy.ALWAYS)
-    private Integer contextLimit;
+    private Long contextLimit;
     @TableField(updateStrategy = FieldStrategy.ALWAYS)
     private Integer streamFirstChunkTimeoutSeconds;
     @TableField(updateStrategy = FieldStrategy.ALWAYS)
